@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <PackAsTool>true</PackAsTool>
 
     <!-- Make sure start same folder .NET Core CLI and Visual Studio -->
@@ -25,8 +25,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Cake.Frosting" Version="1.0.0-rc0002" />
-    <PackageReference Include="Cake.MarkdownToPdf" Version="2.5.2" />
+    <PackageReference Include="Cake.Frosting" Version="4.0.0" />
+    <PackageReference Include="Cake.MarkdownToPdf" Version="3.0.5" />
   </ItemGroup>
 
 </Project>
