@@ -1,4 +1,6 @@
-﻿namespace EveOPreview.Services
+﻿using EveOPreview.View;
+
+namespace EveOPreview.Services
 {
 	public interface IThumbnailManager
 	{
@@ -7,5 +9,9 @@
 
 		void UpdateThumbnailsSize();
 		void UpdateThumbnailFrames();
+
+		IThumbnailView GetClientByTitle(string title);
+		IThumbnailView GetClientByPointer(System.IntPtr ptr);
+		IThumbnailView GetActiveClient();
 	}
 }
